@@ -3,33 +3,33 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.9.11",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.52",
-    "@testing-library/jest-dom": "^5.5.0",
-    "@testing-library/react": "^10.0.2",
-    "@testing-library/user-event": "^10.0.2",
-    "@types/number-to-words": "^1.2.0",
-    "axios": "^0.19.2",
-    "history": "^4.10.1",
-    "html2canvas": "^1.0.0-rc.5",
-    "js-cookie": "^2.2.1",
+    "@material-ui/core": "^4.12.3",
+    "@material-ui/icons": "^4.11.2",
+    "@material-ui/lab": "^4.0.0-alpha.60",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^12.0.0",
+    "@testing-library/user-event": "^13.2.1",
+    "@types/number-to-words": "^1.2.1",
+    "axios": "^0.21.1",
+    "history": "^5.0.1",
+    "html2canvas": "^1.3.2",
+    "js-cookie": "^3.0.0",
     "materialize-css": "^1.0.0",
     "number-to-words": "^1.2.4",
-    "react": "^16.13.1",
-    "react-axios": "^2.0.3",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.4.1",
-    "react-select": "^3.1.0",
+    "react": "^17.0.2",
+    "react-axios": "^2.0.5",
+    "react-dom": "^17.0.2",
+    "react-redux": "^7.2.4",
+    "react-router-dom": "^5.2.0",
+    "react-scripts": "^4.0.3",
+    "react-select": "^4.3.1",
     "react-select-tile": "^1.0.10",
-    "react-tabs": "^3.1.0",
-    "redux": "^4.0.5",
-    "redux-devtools-extension": "^2.13.8",
+    "react-tabs": "^3.2.2",
+    "redux": "^4.1.1",
+    "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.3.0",
     "use-axios-react": "^0.2.3",
-    "victory": "^34.1.3"
+    "victory": "^35.10.1"
   },
   "scripts": {
     "start": "react-scripts start",
