@@ -29,13 +29,15 @@ const App = () => {
     <Provider store={store}>
       <BrowserRouter history={history}>
         <Navbar />
-        <Route exact path='/' component={Landing} />
-        <Route exact path='/checkloggedin' component={CheckLoggedIn} />
-        <SemiPrivateRoute exact path='/helform' component={HelForm} />
-        <PrivateRoute exact path='/Dashboard' component={Dashboard} />
-        <PrivateRoute exact path='/helData' component={HELData} />
-        <PrivateRoute exact path='/create' component={CreateTimeTable} />
-        <PrivateRoute exact path='/aboutUs' component={About} />
+        <Route exact path="/" component={Landing} />
+        <div className="body-wrapper">
+          <Route exact path="/checkloggedin" component={CheckLoggedIn} />
+          <SemiPrivateRoute exact path="/helform" component={HelForm} />
+          <PrivateRoute exact path="/Dashboard" component={Dashboard} />
+          <PrivateRoute exact path="/helData" component={HELData} />
+          <PrivateRoute exact path="/create" component={CreateTimeTable} />
+          <PrivateRoute exact path="/aboutUs" component={About} />
+        </div>
       </BrowserRouter>
     </Provider>
   );
