@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import "react-tabs/style/react-tabs.css";
 import M from "materialize-css";
-import "materialize-css/dist/css/materialize.min.css";
+// import 'materialize-css/dist/css/materialize.min.css';
 import PropTypes from "prop-types";
 import SearchHel from "./SearchHel.jsx";
 import Search from "../utils/Search.jsx";
@@ -15,23 +15,26 @@ class SearchTabs extends Component {
     super(props);
     this.state = {
       initial: this.props.allCourses,
-      current: this.props.allCourses
+      current: this.props.allCourses,
     };
     this.filterItems = this.filterItems.bind(this);
   }
 
   filterItems(input) {
-    let filterCourses = obj =>
+    let filterCourses = (obj) =>
       Object.keys(obj)
         .filter(
-          item =>
+          (item) =>
             item.toLowerCase().search(input.target.value.toLowerCase()) !==
               -1 ||
             obj[item]["name"]
               .toLowerCase()
               .search(input.target.value.toLowerCase()) !== -1
         )
-        .reduce((res, key) => {res[key] = obj[key]; return res}, {});
+        .reduce((res, key) => {
+          res[key] = obj[key];
+          return res;
+        }, {});
     let updatedlist = filterCourses(this.state.initial);
     this.setState({ current: updatedlist });
   }
@@ -39,26 +42,26 @@ class SearchTabs extends Component {
   render() {
     return (
       <>
-        <div className='col s12'>
+        <div className="col s12">
           <ul
-            ref={Tabs => {
+            ref={(Tabs) => {
               this.Tabs = Tabs;
             }}
-            className='tabs tabs-fixed-width tab-demo'
+            className="tabs tabs-fixed-width tab-demo"
           >
-            <li className='tab col s6'>
-              <a href='#ALL'>All Courses</a>
+            <li className="tab col s6">
+              <a href="#ALL">All Courses</a>
             </li>
-            <li className='tab col s6'>
-              <a href='#HEL'>Humanities Electives</a>
+            <li className="tab col s6">
+              <a href="#HEL">Humanities Electives</a>
             </li>
           </ul>
         </div>
-        <div id='ALL' className='col s12'>
+        <div id="ALL" className="col s12">
           <Search action={this.filterItems} />
           <ListCourse courses={this.state.current} />
         </div>
-        <div id='HEL' className='col s12'>
+        <div id="HEL" className="col s12">
           <SearchHel
             currentHels={this.props.currentHels}
             onSelect={this.props.onSelectOption}
@@ -104,7 +107,7 @@ SearchTabs.propTypes = {
   current: PropTypes.string.isRequired,
   onChangeTab: PropTypes.func.isRequired,
   currentHels: PropTypes.object,
-  onSelectOption: PropTypes.func.isRequired
+  onSelectOption: PropTypes.func.isRequired,
 };
 
 export default SearchTabs;
