@@ -12,31 +12,31 @@ class SectionTabs extends Component {
   render() {
     return (
       <>
-        <div className='col s12'>
+        <div className="col s12">
           <ul
-            ref={Tabs => {
+            ref={(Tabs) => {
               this.Tabs = Tabs;
             }}
-            className='tabs tabs-fixed-width tab-demo'
+            className="tabs tabs-fixed-width tab-demo"
           >
-            <li className='tab col s6'>
-              <a href='#Lecture'>Lecture</a>
+            <li className="tab col s6">
+              <a href="#Lecture">Lecture</a>
             </li>
-            <li className='tab col s6'>
-              <a href='#Tutorial'>Tutorial</a>
+            <li className="tab col s6">
+              <a href="#Tutorial">Tutorial</a>
             </li>
-            <li className='tab col s6'>
-              <a href='#Practical'>Practical</a>
+            <li className="tab col s6">
+              <a href="#Practical">Practical</a>
             </li>
           </ul>
         </div>
-        <div id='Lecture' className='col s12'>
+        <div id="Lecture" className="col s12">
           <ListSection sections={this.props.getSections("L")} />
         </div>
-        <div id='Tutorial' className='col s12'>
+        <div id="Tutorial" className="col s12">
           <ListSection sections={this.props.getSections("T")} />
         </div>
-        <div id='Practical' className='col s12'>
+        <div id="Practical" className="col s12">
           <ListSection sections={this.props.getSections("P")} />
         </div>
       </>
@@ -45,7 +45,7 @@ class SectionTabs extends Component {
 }
 
 SectionTabs.propTypes = {
-  getSections: PropTypes.func.isRequired
+  getSections: PropTypes.func.isRequired,
 };
 
 export default SectionTabs;
